+++ src/detectors/python/index.ts
@@ -1,0 +1,3 @@
+export { pythonDeadAbstractionDetector } from "./deadAbstraction.js";
+export { pythonDuplicateLogicDetector } from "./duplicateLogic.js";
+export { pythonTodoCommentDetector } from "./todoComment.js";